Contact Sync reintegrated

+++ Depo/Depo/ContactSyncController.m
@@ -83,12 +83,12 @@
             switch (status.status) {
                 case SYNC_RESULT_SUCCESS: {
                     ContactSyncResult *currentSyncResult = [[ContactSyncResult alloc] init];
-                    currentSyncResult.clientUpdateCount = status.updatedContactsReceived.count;
-                    currentSyncResult.serverUpdateCount = status.updatedContactsSent.count;
-                    currentSyncResult.clientNewCount = status.createdContactsReceived.count;
-                    currentSyncResult.serverNewCount = status.createdContactsSent.count;
-                    currentSyncResult.clientDeleteCount = status.deletedContactsOnDevice.count;
-                    currentSyncResult.serverDeleteCount = status.deletedContactsOnServer.count;
+                    currentSyncResult.clientUpdateCount = (int)status.updatedContactsReceived.count;
+                    currentSyncResult.serverUpdateCount = (int)status.updatedContactsSent.count;
+                    currentSyncResult.clientNewCount = (int)status.createdContactsReceived.count;
+                    currentSyncResult.serverNewCount = (int)status.createdContactsSent.count;
+                    currentSyncResult.clientDeleteCount = (int)status.deletedContactsOnDevice.count;
+                    currentSyncResult.serverDeleteCount = (int)status.deletedContactsOnServer.count;
                     currentSyncResult.syncType = APPDELEGATE.session.syncType;
                     APPDELEGATE.session.syncResult = currentSyncResult;
                     [SyncUtil writeLastContactSyncResult:currentSyncResult];
